Create CNN_train_and_test as a first attempt at making and training the model using only numpy
Imports and utilizes all the other files to train and test a file. Returns an accuracy percentage in the form "Epoch n: Accuracy x". Applies Convolution, Maxpooling, and softmax to make the network. Requires further optimization to allow for quicker training, some leads include numpy vectorization and batch processing. Unsure of current performance as training is still ongoing of the model on the local machine as of uploading